move ServerConfig to subpackage
