+++ common/src/main/java/org/vivecraft/server/ServerNetworking.java
@@ -17,6 +17,7 @@ import org.vivecraft.common.network.CommonNetworkHelper;
 import org.vivecraft.common.network.VrPlayerState;
 import org.vivecraft.common.CommonDataHolder;
 import org.vivecraft.mixin.server.ChunkMapAccessor;
+import org.vivecraft.server.config.ServerConfig;
 
 import java.util.*;
