+++ fabric/src/main/java/org/vivecraft/fabric/VivecraftMod.java
@@ -2,7 +2,7 @@ package org.vivecraft.fabric;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
-import org.vivecraft.server.ServerConfig;
+import org.vivecraft.server.config.ServerConfig;
 import org.vivecraft.server.ServerUtil;
 
 public class VivecraftMod implements ModInitializer {
